Add SSI263.cpp/h and class-ify (WIP)
