+++ AppleWinExpress2008.vcproj
@@ -782,6 +782,14 @@
 					>
 				</File>
 				<File
+					RelativePath=".\source\SSI263.cpp"
+					>
+				</File>
+				<File
+					RelativePath=".\source\SSI263.h"
+					>
+				</File>
+				<File
 					RelativePath=".\source\SynchronousEventManager.cpp"
 					>
 				</File>
